refactor: mainpage refactoring
