+++ client/src/components/Comments.js
@@ -1,11 +1,14 @@
-import React, { useState } from 'react'
+import React, { useState, useContext } from 'react'
 import { AppContext } from "../AppContext";
-
-import { Stack,Button } from '@mui/material';
+import { ThemeProvider } from '@mui/material/styles';
+import { Stack,Button, fabClasses } from '@mui/material';
 import Typography from '@mui/material/Typography';
 import TextField from '@mui/material/TextField';
+import "../utils/Comments.css"
+import { Box } from '@mui/system';
 
 export const Comments = () => {
+  const context = useContext(AppContext);
   const [comment, setComment] = useState('');
   const [id, setId] = useState(1);
   const [commentLists, setCommentLists] = useState([]);
@@ -19,6 +22,7 @@ export const Comments = () => {
     };
     setCommentLists([...commentLists, newCommentLists]);
     setComment('');
+    setIsValid(false);
   };
 
   // const CommentList = () => {
@@ -31,43 +35,40 @@ export const Comments = () => {
   // }
 
   return (
-
-    <Stack >
-      <Stack direction="row" sx={{ marginY:"20px" }} >
-        <TextField 
-          type="text"
-          placeholder="Add a comment…"
-          style = {{width: '100%', color: 'black'}}
-          // inputProps={ariaLabel}
-          onChange={(e) => {
-            setComment(e.target.value);
-          }}
-          onKeyUp={(e) => {
-            e.target.value.length > 0 ? setIsValid(true) : setIsValid(false);
-          }}
-          value={comment}
-        />
-        <Button 
-          type="button"
-          onClick={post}
-          disabled={isValid ? false : true}
-          >
-          post
-        </Button>
+    <ThemeProvider theme={context.state.theme}>
+      <Stack>
+        <Stack className='comments-text' direction="row" sx={{ marginY:"20px", color:"text.secondary" }} >
+          <Box
+            component="input"
+            type="text"
+            placeholder="Add a comment…"
+            sx = {{width: '100%'}}
+            onChange={(e) => {
+              setComment(e.target.value);
+            }}
+            onKeyUp={(e) => {
+              (e.target.value.length > 0) ? setIsValid(true) : setIsValid(false);
+            }}
+            value={comment}
+          />
+          <Button
+            type="button"
+            onClick={post}
+            disabled={isValid ? false : true}
+            >
+            post
+          </Button>
+        </Stack>
+        {commentLists.map((comment) => {
+          return (
+            <Stack direction="row">
+              <Typography color="text.secondary"> {comment.id}</Typography>
+              <Typography>{comment.content}</Typography>
+            </Stack>
+          )
+        })}
       </Stack>
-      
-      {commentLists.map((comment) => {
-        return (
-          <Stack direction="row">
-            <Typography color="text.secondary"> {comment.id}</Typography>
-            <Typography>{comment.content}</Typography>
-          </Stack>
-        )
-      })}
-      </Stack>
-      
-  
-    
+      </ThemeProvider>
   )  
 }
